+++ .gitignore
@@ -1,0 +1,4 @@
+target/
+.idea
+src/main/java/resources/teamsNames.txt
+src/main/java/resources/resultsDetails.txt
