+++ src/main/resources/teamsNames.txt
@@ -1,0 +1,1 @@
+1. AC Omonia Nicosia<br>2. AS Monaco<br>3. Benfica Lisbon<br>4. Bodoe/Glimt<br>5. CS Fola Esch<br>6. Celtic Glasgow<br>7. Connah`s Quay Nomads FC<br>8. FC Alashkert Yerevan<br>9. FC CFR 1907 Cluj<br>10. FC Dinamo Tbilisi<br>11. FC Kairat Almaty<br>12. FC Midtjylland<br>13. FC Prishtina<br>14. FC Shakhtar Donetsk<br>15. FC Shakhter Soligorsk<br>16. FC Sheriff Tiraspol<br>17. FK Borac Banja Luka<br>18. FK Buducnost<br>19. FK Crvena Zvezda Belgrade<br>20. FK Spartak Moscow<br>21. Ferencvarosi Budapest<br>22. Flora Tallinn<br>23. GNK Dinamo Zagreb<br>24. Galatasaray Istanbul<br>25. Glasgow Rangers<br>26. HB Torshavn<br>27. HJK Helsinki<br>28. Hibernians FC Paola<br>29. Inter Club de Escaldes<br>30. KF Teuta<br>31. KRC Genk<br>32. Legia Warszawa<br>33. Lincoln Red Imps<br>34. Linfield FC<br>35. Maccabi Haifa FC<br>36. Malmo FF<br>37. Mura Murska Sobota<br>38. Neftchi Baku PFC<br>39. Olympiacos Piraeus<br>40. PFC Ludogorets 1945 Razgrad<br>41. PSV Eindhoven<br>42. Riga FC<br>43. SK Rapid Wien<br>44. SS Folgore Falciano Calcio<br>45. Shamrock Rovers<br>46. Shkendija Tetovo<br>47. Slavia Prague<br>48. Slovan Bratislava<br>49. Sparta Prague<br>50. Valur Reykjavik<br>51. Vilnius FK Zalgiris<br>52. Young Boys Bern<br>
